Improvements...
I made some improvements to the site, but I'm not really finished with
them yet.
- First off I changed the menu to select a planet to a completely
different screen where you can pick a planet. On this page there is also
a little bit of information about the site (well, there will b ein the
future), and links to share the page on social media (Facebook, Twitter
and Reddit)
- Secondly the text color of the input in the intro screen will now turn
red if the input value is invalid.
- And lastly I already changed the organization of the resources dir,
next update there will be thumbnails (of the planets) and icons (favicon
and icons for the share buttons).