Switched back to a list :hankey:
I decided to remove the grid, I couldn't come up with a nice design for
it, so switched back to a menu. Not the menu there was a few commits
back, but a totally new design. :tada:
Because of that I could remove all the thumbnail images and switch them
to small icons (24x24).

Furthermore I started improving the CSS file...