Code update & new features
I'll start with the new features and new things. First of all, the page
now has a favicon (rescources\icons\favicon.png).
Also, I added to icons to the menu page. You can click on them to select
the text that will show up. The I is for information about the page,
this is the text that was already there. The erlenmeyer flask will give
you some information about the science behind this page (It explains
that your mass isn't actually change, what is changing is the force).
And last but certainly not least, along with the latest update to the
InterstellarScaler.js UDF, you can now give a weight on every planet and
the application will tell you what you weigh on other planets in
relation to that specific planet.

And then the updates to the code. I not only tried to improve the code,
but I also tried to make the code more elegant. Especially the HTML and
JavaScript should be much better now, since I already focussed on CSS in
the previous commit.