Added the imperial system
The application will now try to get the country of the user and if it is
the US it will change its system of measurement to imperial (pounds)
from metric (kilograms). Also, you can now find a button to manually
switch from imperial to metric.

Besides that, I found out that the image I used for mars actually was an
image of the moon... :confounded: So I updated it :smile: